requirements.txt
# Primary deps
click>=7.0, <8.0
gitpython>=3.0.5, <4
jinja2>=2.11.1, <3
python-slugify>=4.0.0, <5
pyyaml>=5.3, <6

# Testing deps
deepdiff>=4.2.0, <5
pyfakefs>=3.7, <4
pytest>=5.3.4, <6
pytest-cov>=2.8.1, <3
pytest-mock>=2.0.0, <3
